Added encryption key generation fix

+++ encryption.py
@@ -1,44 +1,39 @@
 from cryptography.fernet import Fernet
+import os
 
 
 def generate_key():
+    if not os.path.exists("key.key"):
+        key = Fernet.generate_key()
 
-    key = Fernet.generate_key()
-
-    with open("key.key", "wb") as file:
-        file.write(key)
-
+        with open("key.key", "wb") as file:
+            file.write(key)
 
 
 def load_key():
+    generate_key()
 
     with open("key.key", "rb") as file:
         return file.read()
 
 
-
 def encrypt_password(password):
-
     key = load_key()
 
-    cipher = Fernet(key)
+    fernet = Fernet(key)
 
-    encrypted_password = cipher.encrypt(
-        password.encode()
-    )
+    encrypted = fernet.encrypt(password.encode())
 
-    return encrypted_password
-
+    return encrypted
 
 
 def decrypt_password(encrypted_password):
-
     key = load_key()
 
-    cipher = Fernet(key)
+    fernet = Fernet(key)
 
-    decrypted_password = cipher.decrypt(
-        encrypted_password
+    decrypted = fernet.decrypt(
+        encrypted_password.encode()
     )
 
-    return decrypted_password.decode()
+    return decrypted.decode()
